requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
json5==0.9.14
lxml==4.9.3
fake-useragent==1.4.0
selenium==4.15.2
webdriver-manager==4.0.1
pandas==2.1.3
numpy==1.24.3
openrouter==1.0
PyPDF2==3.0.1 